// game.server/src/Generated/packet/room_actions.ts
export interface C2SCreateRoomRequest {
	/**
	 * @generated from protobuf field: string name = 1
	 */
	name: string;
	/**
	 * @generated from protobuf field: int32 maxUserNum = 2
	 */
	maxUserNum: number;
}
// ...
class C2SCreateRoomRequest$Type extends MessageType<C2SCreateRoomRequest> {
	constructor() {
		super('C2SCreateRoomRequest', [
			{ no: 1, name: 'name', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
			{ no: 2, name: 'maxUserNum', kind: 'scalar', T: 5 /*ScalarType.INT32*/ },
		]);
	}
	create(value?: PartialMessage<C2SCreateRoomRequest>): C2SCreateRoomRequest {
		const message = globalThis.Object.create(this.messagePrototype!);
		message.name = '';
		message.maxUserNum = 0;
		if (value !== undefined) reflectionMergePartial<C2SCreateRoomRequest>(this, message, value);
		return message;
	}
	internalBinaryRead(
		reader: IBinaryReader,
		length: number,
		options: BinaryReadOptions,
		target?: C2SCreateRoomRequest,
	): C2SCreateRoomRequest {
		let message = target ?? this.create(),
			end = reader.pos + length;
		while (reader.pos < end) {
			let [fieldNo, wireType] = reader.tag();
			switch (fieldNo) {
				case /* string name */ 1:
					message.name = reader.string();
					break;
				case /* int32 maxUserNum */ 2:
					message.maxUserNum = reader.int32();
					break;
				default:
					let u = options.readUnknownField;
					if (u === 'throw')
						throw new globalThis.Error(
							`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
						);
					let d = reader.skip(wireType);
					if (u !== false)
						(u === true ? UnknownFieldHandler.onRead : u)(
							this.typeName,
							message,
							fieldNo,
							wireType,
							d,
						);
			}
		}
		return message;
	}
	internalBinaryWrite(
		message: C2SCreateRoomRequest,
		writer: IBinaryWriter,
		options: BinaryWriteOptions,
	): IBinaryWriter {
		/* string name = 1; */
		if (message.name !== '') writer.tag(1, WireType.LengthDelimited).string(message.name);
		/* int32 maxUserNum = 2; */
		if (message.maxUserNum !== 0) writer.tag(2, WireType.Varint).int32(message.maxUserNum);
		let u = options.writeUnknownFields;
		if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
		return writer;
	}
}
/**
 * @generated MessageType for protobuf message C2SCreateRoomRequest
 */
export const C2SCreateRoomRequest = new C2SCreateRoomRequest$Type();